The client and server are now working to correctly get data from the database from the rental_price_info table and update state.
The general structure for the client and server to get data from rental_availability is mostly correct, but I am getting a 404 error on the client side. Will need to investigate further.

During the course of working with the server and frontend I realized that there are two issues with my database:
1. The availability should be either 0 or 1, but is currently only 0.
2. The dates were not entered into the db correctly.

I will update the seed and schema to resolve these issues and then push with a solution.